+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Unreleased changes are available as `coupergateway/couper:edge` container.
 
 * **Added**
+  * `argon2i` and `argon2id` password hash support for [`basic_auth`](https://docs.couper.io/configuration/block/basic_auth) `htpasswd_file` ([#866](https://github.com/coupergateway/couper/issues/866))
   * [`beta_rate_limiter` access control](https://docs.couper.io/configuration/block/rate_limiter) for fixed/sliding window rate limiting ([#881](https://github.com/coupergateway/couper/pull/881))
   * [`beta_introspection` block](https://docs.couper.io/configuration/block/introspection) in [`jwt` block](https://docs.couper.io/configuration/block/jwt) to facilitate token introspection in order to detect revoked tokens ([#649](https://github.com/avenga/couper/pull/649))
   * `startup_delay` attribute for [`job` block](https://docs.couper.io/configuration/block/job) to delay the first job execution after startup ([#836](https://github.com/coupergateway/couper/issues/836))